Style + split global header into global and defs

+++ src/utils.h
@@ -1,6 +1,8 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-unsigned int cur_time_millis(void);
+#include "defs.h"
+
+u64 cur_time_millis(void);
 
 #endif /* UTILS_H */
